sr(api): move training-repository in devcomp
