+++ api/lib/infrastructure/repositories/index.js
@@ -1,7 +1,7 @@
 import { injectDependencies } from '../../../src/shared/infrastructure/utils/dependency-injection.js';
 
 import * as correctionRepository from './correction-repository.js';
-import * as trainingRepository from './training-repository.js';
+import * as trainingRepository from '../../../src/devcomp/infrastructure/repositories/training-repository.js';
 import * as trainingTriggerRepository from './training-trigger-repository.js';
 import * as tutorialEvaluationRepository from './tutorial-evaluation-repository.js';
 import * as tutorialRepository from './tutorial-repository.js';
