+++ api/src/devcomp/infrastructure/repositories/training-repository.js
@@ -1,14 +1,14 @@
-import { Training } from '../../domain/models/Training.js';
-import { TrainingSummary } from '../../domain/read-models/TrainingSummary.js';
-import { knex } from '../../../db/knex-database-connection.js';
-import { NotFoundError } from '../../domain/errors.js';
-import { DomainTransaction } from '../DomainTransaction.js';
-import { UserRecommendedTraining } from '../../domain/read-models/UserRecommendedTraining.js';
-import { fetchPage } from '../utils/knex-utils.js';
+import { Training } from '../../../../lib/domain/models/Training.js';
+import { TrainingSummary } from '../../../../lib/domain/read-models/TrainingSummary.js';
+import { knex } from '../../../../db/knex-database-connection.js';
+import { NotFoundError } from '../../../../lib/domain/errors.js';
+import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
+import { UserRecommendedTraining } from '../../../../lib/domain/read-models/UserRecommendedTraining.js';
+import { fetchPage } from '../../../../lib/infrastructure/utils/knex-utils.js';
 import lodash from 'lodash';
-import * as trainingTriggerRepository from './training-trigger-repository.js';
-import { TrainingForAdmin } from '../../domain/read-models/TrainingForAdmin.js';
-import { TrainingTrigger } from '../../domain/models/index.js';
+import * as trainingTriggerRepository from '../../../../lib/infrastructure/repositories/training-trigger-repository.js';
+import { TrainingForAdmin } from '../../../../lib/domain/read-models/TrainingForAdmin.js';
+import { TrainingTrigger } from '../../../../lib/domain/models/index.js';
 
 const { pick } = lodash;
